[155917] NPE if delete server before publishing

+++ plugins/org.eclipse.wst.server.core/servercore/org/eclipse/wst/server/core/internal/PublishInfo.java
@@ -121,7 +121,8 @@ public class PublishInfo {
 			serverIdToPublishInfo.remove(serverId);
 			save();
 		}
-		path.append(path2).toFile().delete();
+		if (path2 != null)
+			path.append(path2).toFile().delete();
 	}
 
 	/**
